msm: reorder init calls in board's init_machine handler
There are a few minor ordering issues that cause us to have either extra
code in drivers or issues with doing proper memory mapping in drivers for
regions of the memory map they use.

We should be populating the device structures from DT before we call any
particular driver init so that the drivers probe function can be called
at the same time (with a proper platform device).

Additionally, we should be calling smem_init before socinfo_init as
socinfo utilizes smem.  (We have extra code in smem to handle calls before
smem_init/smem_probe are called that we could remove if we order init
calls properly).

Change-Id: If616190f7310b7779e0ff1702bb0f873f4c5471d
Signed-off-by: Kumar Gala <galak@codeaurora.org>
Signed-off-by: Zdrowy Gosciu <ZdrowyGosciu+GITHUB@gmail.com>

Conflicts:
	arch/arm/mach-msm/oppo/board-8974-oppo.c