audit: Increase rate-limit and backlog-limit
Device is sometimes crashing at boot (and at other times) due to audit rate-limit being exceeded.

Forcefully increase the rate-limit and backlog-limit to fix this. There should be a proper way to change this from userspace, but I don't know what that method is. It should also be possible to simply address the denials, but the rate-limit as it is seems very small anyway.

Signed-off-by: Sultanxda <sultanxda@gmail.com>
Signed-off-by: Zdrowy Gosciu <ZdrowyGosciu+GITHUB@gmail.com>