msm: thermal: simple: Introduce standalone MSM thermal solution
Based off of my custom thermal driver for MSM8x60.

This enables thermal monitoring capability in the kernel in the absence of a system wide thermal monitoring entity.
Monitors TSENS temperature and limits the max frequency of the cores. The code is simple, uses 3 throttling steps,
and is completely configurable via sysfs.

All thermal zones must be configured before enabling thermal mitigation. Example for configuring
one of the 3 thermal zones (high, mid, low) :
echo "1728000 52 48" > /sys/kernel/msm_thermal/low_thresh

Where 1728000 is the frequency to throttle to in KHz, 52 is the throttle temperature in degrees C,
and 48 is the unthrottle temperature in degrees C. None of the temperatures set for a zone are allowed to overlap
the temperatures set for another zone. Once all zones all configured, do this to enable thermal mitigation :
echo "1" > /sys/kernel/msm_thermal/enabled

Update sysfs node :
There are times when we want to modify the maxfreq from userspace (i.e. low-power mode), so add an interface
for that as the thermal driver is formally in control of the maxfreq.
Prevents maxfreq overrides from a certain closed-source userspace daemon (mpdecision), as well as from anything else in userland.
Cannot override a thermal throttle if the throttle maxfreq is less than the requested user maxfreq.
Usage : write the desired maxfreq to /sys/kernel/msm_thermal/user_maxfreq; write '0' to disable the maxfreq override.

Commit contains :
* Initial commit https://github.com/sultanxda/android_kernel_oneplus_msm8974/commit/b24edc5aeea5c61d89f747c3e3650d8b5efc8d5d
* Fix on/off switch https://github.com/sultanxda/android_kernel_oneplus_msm8974/commit/30eb2bbaff3fa5e184bb31a7f8f440e785690d26
* Clean up 'enabled' sysfs node https://github.com/sultanxda/android_kernel_oneplus_msm8974/commit/0b0c85f8aac51af7a78d72bc67cfe69534a6d574
* Remove unnecessary check https://github.com/sultanxda/android_kernel_oneplus_msm8974/commit/ccfbfb3520d73e8ab3c63a70586d13c428f003a3
* Add sysfs node to manually control maxfreq https://github.com/sultanxda/android_kernel_oneplus_msm8974/commit/36abb99bdaefb231bee3b2b1bad3137a28df7043

Signed-off-by: Sultanxda <sultanxda@gmail.com>
Signed-off-by: Zachariah <zkennedy87@gmail.com>
Signed-off-by: Zdrowy Gosciu <ZdrowyGosciu+GITHUB@gmail.com>